# apps/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "0x1audio",
    "summary": "Private music catalog with search, chat, and playlists.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/graph": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Get Graph",
        "operationId": "getGraph",
        "parameters": [
          {
            "name": "focus_track_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Focus Track Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 24,
              "minimum": 2,
              "default": 12,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/playlists": {
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "List Playlists",
        "operationId": "listPlaylists",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaylistSummaryOut"
                  },
                  "type": "array",
                  "title": "Response Listplaylists"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Create Playlist",
        "operationId": "createPlaylist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaylistBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playlists/{playlist_id}": {
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "Get Playlist",
        "operationId": "getPlaylist",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Playlist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "playlists"
        ],
        "summary": "Update Playlist",
        "operationId": "updatePlaylist",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Playlist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaylistBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "playlists"
        ],
        "summary": "Delete Playlist",
        "operationId": "deletePlaylist",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Playlist Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playlists/{playlist_id}/tracks": {
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Add Playlist Tracks",
        "operationId": "addPlaylistTracks",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Playlist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaylistTracksBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "playlists"
        ],
        "summary": "Remove Playlist Tracks",
        "operationId": "removePlaylistTracks",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Playlist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemovePlaylistTracksBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playlists/{playlist_id}/tracks/order": {
      "put": {
        "tags": [
          "playlists"
        ],
        "summary": "Reorder Playlist Tracks",
        "operationId": "reorderPlaylistTracks",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Playlist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderPlaylistBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistRefOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ArtistRefOut"
      },
...
      "CreatePlaylistBody": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "color": {
            "$ref": "#/components/schemas/PlaylistColor"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "track_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Ids"
          }
        },
        "type": "object",
        "required": [
          "title",
          "color"
        ],
        "title": "CreatePlaylistBody"
      },
...
      "GraphLinkOut": {
        "properties": {
          "source": {
            "type": "string",
            "format": "uuid",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "format": "uuid",
            "title": "Target"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "audio_weight": {
            "type": "number",
            "title": "Audio Weight"
          },
          "profile_weight": {
            "type": "number",
            "title": "Profile Weight"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "weight",
          "audio_weight",
          "profile_weight",
          "reasons"
        ],
        "title": "GraphLinkOut"
      },
      "GraphNodeOut": {
        "properties": {
          "track": {
            "$ref": "#/components/schemas/TrackOut"
          }
        },
        "type": "object",
        "required": [
          "track"
        ],
        "title": "GraphNodeOut"
      },
      "GraphResponseOut": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeOut"
            },
            "type": "array",
            "title": "Nodes"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/GraphLinkOut"
            },
            "type": "array",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "links"
        ],
        "title": "GraphResponseOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaylistColor": {
        "type": "string",
        "enum": [
          "light_blue",
          "washed_blue",
          "dark_blue",
          "teal",
          "mint",
          "green",
          "forest",
          "yellow",
          "amber",
          "orange",
          "coral",
          "light_red",
          "rose",
          "dark_red",
          "pink",
          "magenta",
          "lavender",
          "purple",
          "violet",
          "indigo",
          "slate",
          "charcoal"
        ],
        "title": "PlaylistColor"
      },
      "PlaylistDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaylistColor"
              },
              {
                "type": "null"
              }
            ]
          },
          "theme_colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Theme Colors",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "type": "array",
            "title": "Tracks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "created_at",
          "updated_at",
          "tracks"
        ],
        "title": "PlaylistDetailOut"
      },
      "PlaylistSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaylistColor"
              },
              {
                "type": "null"
              }
            ]
          },
          "theme_colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Theme Colors",
            "default": []
          },
          "track_count": {
            "type": "integer",
            "title": "Track Count"
          },
          "cover_colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cover Colors",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "track_count",
          "created_at",
          "updated_at"
        ],
        "title": "PlaylistSummaryOut"
      },
      "PlaylistTracksBody": {
        "properties": {
          "track_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Track Ids"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "track_ids"
        ],
        "title": "PlaylistTracksBody"
      },
...
      "RemovePlaylistTracksBody": {
        "properties": {
          "track_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Track Ids"
          }
        },
        "type": "object",
        "required": [
          "track_ids"
        ],
        "title": "RemovePlaylistTracksBody"
      },
      "ReorderPlaylistBody": {
        "properties": {
          "track_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Track Ids"
          }
        },
        "type": "object",
        "required": [
          "track_ids"
        ],
        "title": "ReorderPlaylistBody"
      },
...
      "TrackOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/TrackStatus"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          },
          "indexed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "cover_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Color"
          },
          "has_cover": {
            "type": "boolean",
            "title": "Has Cover",
            "default": false
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "is_instrumental": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Instrumental"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistRefOut"
            },
            "type": "array",
            "title": "Artists",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "artist",
          "original_filename",
          "content_type",
          "size_bytes",
          "status",
          "error_message",
          "imported_at",
          "indexed_at"
        ],
        "title": "TrackOut"
      },
      "TrackStatus": {
        "type": "string",
        "enum": [
          "uploading",
          "queued",
          "indexing",
          "ready",
          "failed"
        ],
        "title": "TrackStatus"
      },
      "UpdatePlaylistBody": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "clear_description": {
            "type": "boolean",
            "title": "Clear Description",
            "default": false
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaylistColor"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "UpdatePlaylistBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}